Fix: 빌드 오류 수정(suspense 추가)
